+++ swift-sdk/Internal/InternalIterableAPI.swift
@@ -1152,7 +1152,7 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
         }
     }
     
-    func getCriteriaData(completion: @escaping (Data) -> Void) {
+    func getCriteriaData(completion: @escaping (Data) -> Void, onFailure: @escaping (String) -> Void) {
         apiClient.getCriteria().onSuccess { data in
             do {
                 let jsonData = try JSONSerialization.data(withJSONObject: data, options: [])
@@ -1160,8 +1160,10 @@ final class InternalIterableAPI: NSObject, PushTrackerProtocol, AuthProvider {
             } catch {
                 print("Error converting dictionary to data: \(error)")
             }
+        }.onError { error in
+            onFailure(error.reason ?? error.localizedDescription)
         }
-	}
+    }
 
     private func createDefaultMobileFrameworkInfo() -> IterableAPIMobileFrameworkInfo {
         let frameworkType = IterableAPIMobileFrameworkDetector.frameworkType()